Core accounts module test refactor
Now using the graphql-modules testkit util for testing
Core Module.
